requirements.txt
anthropic>=0.25.0
python-dotenv>=1.0.0
pandas>=2.0.0
jsonlines>=4.0.0
pytest>=8.0.0
tqdm>=4.66.0
numpy>=1.26.0
